Moved each class/interface into its own file
