+++ src/Agent.Sdk/Knob/IKnobSource.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Agent.Sdk.Knob
+{
+
+    public interface IKnobSource
+    {
+        KnobValue GetValue(IKnobValueContext context);
+        string GetDisplayString();
+    }
+
+}
