+++ src/Agent.Worker/Handlers/NodeHandler.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using Agent.Sdk;
+using Agent.Sdk.Knob;
 using Microsoft.VisualStudio.Services.Agent.Util;
 using System.IO;
 using System.Text;
